refactor: refactor context
