+++ src/main/java/com/flav/trailers/context/trailers/users/domain/constants/UserConstants.java
@@ -1,0 +1,8 @@
+package com.flav.trailers.context.trailers.users.domain.constants;
+
+public class UserConstants {
+
+    public static final String USER_NOT_FOUND = "User not found by id: %s";
+    public static final String USER_EXITS = "User name (%s) already exist";
+
+}
