+++ src/main/java/com/flav/trailers/context/trailers/users/domain/repositories/IUserCRUDRepository.java
@@ -1,0 +1,18 @@
+package com.flav.trailers.context.trailers.users.domain.repositories;
+
+import com.flav.trailers.context.trailers.users.domain.models.User;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface IUserCRUDRepository {
+
+    public List<User> findAll();
+    public boolean findByEmail(String email);
+    public Optional<User> findById(Long id);
+    public User create(User user);
+    public User update(User user);
+    public boolean delete(Long id);
+    public Optional<User> findUserforfavorite(Long id);
+
+}
