+++ src/main/java/com/flav/trailers/context/trailers/users/infrastructure/persistence/JPA/IJpaUserRepository.java
@@ -1,0 +1,10 @@
+package com.flav.trailers.context.trailers.users.infrastructure.persistence.JPA;
+
+import com.flav.trailers.context.trailers.users.infrastructure.persistence.entities.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IJpaUserRepository extends JpaRepository<UserEntity, Long> {
+
+    public boolean existsByEmail(String email);
+
+}
